src/requirements.txt.txt
pytorch=1.2.0
torchvision=0.4.0
numpy=1.15.4
tensorboardx=1.4
matplotlib=3.0.1